parser-python/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
pdfplumber==0.10.3
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
